Remove unused unified registry code
